Unit test added
